+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-05-06 23:10:31
-// Commit: perf(core): memoize expensive computations in core
+// Updated at: 2026-05-09 19:52:55
+// Commit: perf(core): optimize memory consumption in core list viewport
 
 class coreComponentState {
-  final String commitRef = '961';
-  final String lastMessage = 'perf(core): memoize expensive computations in core';
-  final DateTime timestamp = DateTime.parse('2026-05-06T23:10:31+05:00');
+  final String commitRef = '979';
+  final String lastMessage = 'perf(core): optimize memory consumption in core list viewport';
+  final DateTime timestamp = DateTime.parse('2026-05-09T19:52:55+05:00');
 }